feat(point-of-sale): Add weight-based product quantity modal with decimal support
- Add `is_weight_unit` boolean field to units table via migration
- Create WeightUnitsSeeder to populate weight unit types (KG, LB, GR, etc.)
- Change quantity columns to decimal type for precise weight measurements
- Implement weight quantity modal in PointOfSale component with validation
- Add modal UI component for entering exact quantities before cart addition
- Update Product and Unit models with weight unit detection logic
- Modify addToCart method to intercept weight-based products and show modal
- Add design specifications and requirements documentation for weight modal feature
- Update POS receipt view to display decimal quantities correctly
- Enhance reports (Commissions, ProductsSold) to handle decimal quantities
- Update seeders to work with new decimal quantity schema
- Add unit tests for weight unit functionality
- Improve UX for fruver and weight-based product businesses by automating quantity entry

+++ app/Http/Controllers/ReportExportController.php
@@ -117,14 +117,14 @@ class ReportExportController extends Controller
             $commission = $this->calculateCommission($item);
             $userCommissions[$userId]['commission'] += $commission;
             $userCommissions[$userId]['sales'] += (float) $item->total;
-            $userCommissions[$userId]['items_count'] += (int) $item->quantity;
+            $userCommissions[$userId]['items_count'] += (float) $item->quantity;
             $userCommissions[$userId]['items'][] = [
                 'invoice_number' => $item->invoice_number,
                 'date' => Carbon::parse($item->sale_date)->format('d/m/Y'),
                 'product_name' => $item->product_name,
                 'category' => $item->category_name,
                 'brand' => $item->brand_name,
-                'quantity' => (int) $item->quantity,
+                'quantity' => (float) $item->quantity,
                 'unit_price' => (float) $item->unit_price,
                 'total' => (float) $item->total,
                 'commission' => $commission,
@@ -185,7 +185,7 @@ class ReportExportController extends Controller
         }
 
         $basePrice = (float) $item->unit_price;
-        $quantity = (int) $item->quantity;
+        $quantity = (float) $item->quantity;
         $commissionValue = (float) $item->commission_value;
         $commissionType = $item->commission_type;
